Implemented snooze, scheduled, and completed buttons for follow up care alert bar

+++ web/src/org/openmrs/module/lancearmstrong/web/controller/FollowupCareFormController.java
@@ -127,7 +127,7 @@ public class FollowupCareFormController extends SimpleFormController {
                 	log.debug("Nothing is updated. command=" + command);
                 }
             } else if(command != null && command.startsWith("Delete")) {
-                if(id >= 0) {
+                if( id != null && id >= 0) {
                 	LafUtil.getService().getReminderDao().deleteLafReminder(reminders.get(id));                
                 	log.debug("Reminder deleted: " + reminders.get(id).getFollowProcedure().getName()+"/"+reminders.get(id).getCompleteDate() + "/" + reminders.get(id).getResponseType() + "/" + reminders.get(id).getResponseComments());
                 	request.getSession().setAttribute(WebConstants.OPENMRS_MSG_ATTR, reminders.get(id).getFollowProcedure().getName()+"/"+reminders.get(id).getCompleteDate() + " " + reminders.get(id).getResponseType());
